Added a prober for MacRoman encoding.
MacRoman is not in particularly common use anymore, as it has been
deprecated by Mac OS for over a decade. However, there are programs such
as Microsoft Office for Mac that didn't get the memo, and will output in
MacRoman by default.

The MacRoman detector works similarly to the Latin-1 detector, but
starts at a lower probability.