Add puzzlet's copy of chardet.
This now works on python 2 & 3 and has ~ 130 fewer test failures. Working
backwards indeed proved helpful. It's just not where I want it to be (yet)
with flake8 compatibility for future maintainability. As soon as it is there,
I'll move it over to charade/ and petition @kennethreitz for the ability to
release the package.