Add C-equivalent implementation of filter_english_letters.
Much like "filterWithoutEnglishLetters", it doesn't actually strip out
non-English characters. Instead, it's a filter that is used by
Latin1Prober to inaccurately remove HTML tags.  In reality, it removes
a lot of other stuff too, but I'm not sure if that was intentional yet.

Also cleaned up some of the pylint warnings for charsetprober.py.