Fix chardet.detect exception
First, it is wrong: ValueError is about bad value of the right type,
but the error and the check is about type, not value.

Second, it lies about argument type assuming it unicode for all the
cases that are not valid. What gives you happy debug times until you
finally open chardet source code.

Third, chardet need to respect bytearray type as it could work with it
fine and which one is a legit for holding binary strings in PY3.