Look specifically for "poison" strings
This attempts to address some of @sigmavirus24's concerns over
the utility of this test: Instead of merely looking for strings
where chardet can't find an encoding, we look for prefixes that
prevent any string (or at least every string we checked) that
starts with that prefix from having a detected encoding.

This makes the test significantly more complicated, but should lead
to more robust examples.