Cache hypothesis storage directory
This isn't strictly necessary, but the new test is quite
slow to minimize when a failure is found, which it typically
is right now. Caching the .hypothesis directory means that
examples will persist between runs, which both speeds up the
build and helps keep consistency between runs so that causes
to failures persist.